# docs/05_API/openapi.yaml
openapi: 3.1.0
info:
  title: RecoBridge Recommendation API
  version: 1.0.0
  description: REST contract for recommendation serving and feedback ingestion.
servers:
  - url: http://localhost:8000
security:
  - bearerAuth: []
paths:
  /v1/recommendations:
    post:
      operationId: createRecommendations
      summary: Return top-N recommendations
      parameters:
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendationRequest'
      responses:
        '200':
          description: Recommendation response, possibly degraded via fallback
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/UnavailableError'
  /v1/recommendations/related:
    post:
      operationId: createRelatedRecommendations
      summary: Return products related to a seed product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RelatedRequest'
      responses:
        '200':
          description: Related items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationResponse'
  /v1/events/exposure:
    post:
      operationId: createExposureEvent
      summary: Record rendered recommendation items
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExposureEvent'
      responses:
        '200':
          description: Durably committed or deduplicated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventAccepted'
        '409':
          description: Same idempotency key with different payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/events/feedback:
    post:
      operationId: createFeedbackEvent
      summary: Record click, add-to-cart, remove-from-cart, or purchase feedback
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackEvent'
      responses:
        '200':
          description: Durably committed or deduplicated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventAccepted'
  /v1/health/live:
    get:
      security: []
      operationId: getLiveness
      responses:
        '200':
          description: Process is alive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /v1/health/ready:
    get:
      security: []
      operationId: getReadiness
      responses:
        '200':
          description: Service is ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          $ref: '#/components/responses/UnavailableError'
  /v1/models/version:
    get:
      operationId: getModelVersion
      responses:
        '200':
          description: Active versions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelVersionResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      schema:
        type: string
        maxLength: 128
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnavailableError:
      description: Service or required dependency unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    RecommendationContext:
      type: object
      additionalProperties: false
      required: [page_type]
      properties:
        page_type:
          type: string
          enum: [home, category, product_detail, cart, checkout]
        product_id:
          type: [string, 'null']
        category_id:
          type: [string, 'null']
        device_type:
          type: [string, 'null']
          enum: [desktop, mobile, tablet, null]
    RecommendationRequest:
      type: object
      additionalProperties: false
      required: [session_id, context]
      properties:
        user_id:
          type: [string, 'null']
          maxLength: 128
        session_id:
          type: string
          minLength: 1
          maxLength: 128
        context:
          $ref: '#/components/schemas/RecommendationContext'
        top_k:
          type: integer
          minimum: 1
          maximum: 50
          default: 12
        strategy:
          type: string
          enum: [hybrid, xgboost, cluster, popular]
          default: hybrid
    RelatedRequest:
      type: object
      additionalProperties: false
      required: [product_id]
      properties:
        product_id:
          type: string
        top_k:
          type: integer
          minimum: 1
          maximum: 50
          default: 12
    RecommendationItem:
      type: object
      additionalProperties: false
      required: [product_id, score, rank]
      properties:
        product_id:
          type: string
        category_id:
          type: [string, 'null']
        price_bucket:
          type: [integer, 'null']
          minimum: 0
        score:
          type: number
        rank:
          type: integer
          minimum: 1
        reason_code:
          type: [string, 'null']
    RecommendationResponse:
      type: object
      additionalProperties: false
      required: [request_id, model_version, feature_version, strategy_used, degraded, ranker_promoted, personalization_source, session_signal_count, items, latency_ms]
      properties:
        request_id:
          type: string
        model_version:
          type: string
        feature_version:
          type: string
        strategy_used:
          type: string
        degraded:
          type: boolean
        ranker_promoted:
          type: boolean
          description: True only when the trained ranker is approved and serving.
        personalization_source:
          type: string
          enum: [session_feedback, historical_category_affinity, recent_popular, item_context]
        session_signal_count:
          type: integer
          minimum: 0
        dominant_category_id:
          type: [string, 'null']
        items:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/RecommendationItem'
        latency_ms:
          type: integer
          minimum: 0
    ExposureItem:
      type: object
      additionalProperties: false
      required: [product_id, position]
      properties:
        product_id:
          type: string
        position:
          type: integer
          minimum: 1
    ExposureEvent:
      type: object
      additionalProperties: false
      required: [request_id, session_id, widget_id, page_type, occurred_at, items]
      properties:
        request_id:
          type: string
        user_id:
          type: [string, 'null']
        session_id:
          type: string
        widget_id:
          type: string
        page_type:
          type: string
        occurred_at:
          type: string
          format: date-time
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ExposureItem'
    FeedbackEvent:
      type: object
      additionalProperties: false
      required: [request_id, session_id, product_id, event_type, occurred_at]
      properties:
        request_id:
          type: string
        user_id:
          type: [string, 'null']
        session_id:
          type: string
        product_id:
          type: string
        event_type:
          type: string
          enum: [click, add_to_cart, remove_from_cart, purchase]
        occurred_at:
          type: string
          format: date-time
    EventAccepted:
      type: object
      additionalProperties: false
      required: [event_id, accepted, duplicate]
      properties:
        event_id:
          type: string
        accepted:
          type: boolean
        duplicate:
          type: boolean
    ErrorDetail:
      type: object
      additionalProperties: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id, retryable]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
            retryable:
              type: boolean
            details:
              type: array
              items:
                $ref: '#/components/schemas/ErrorDetail'
    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok, degraded, unavailable]
        checks:
          type: object
          additionalProperties:
            type: string
    ModelVersionResponse:
      type: object
      required: [model_version, feature_version, strategy_version]
      properties:
        model_version:
          type: string
        feature_version:
          type: string
        strategy_version:
          type: string